backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
numpy==2.2.4
torch==2.6.0
loguru
tqdm
transformers
faiss-cpu
sentence-transformers
beautifulsoup4
pyright
ruff
